CLI_Cloud_Platform/requirements.txt
grpcio>=1.68.0
grpcio-tools>=1.68.0
protobuf>=5.28.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
alembic>=1.13.1
psutil
campay
requests